Refactor DomainModelEditor file having now one file per component
