+++ public_src/js/index.js
@@ -10,7 +10,7 @@ var Bootstrap = require('bootstrap');
 var ScenarioLoader = require('./react/scenarioloader');
 var FragmentEditorComponent = require('./react/fragmenteditor/fragmenteditor');
 var ScenarioEditorComponent = require('./react/scenarioeditor/scenarioeditor');
-var DomainModelEditorComponent = require('./react/domainmodeleditor/domainmodeleditor');
+var DomainModelEditorComponent = require('./react/domainmodeleditor/DomainModelEditor');
 var OLCEditorComponent = require('./react/olceditor');
 var IndexComponent = require('./react/indexpage');
 var Router = require('react-router').Router;
