+++ public_src/js/react/domainmodeleditor/CreateNewClass.js
@@ -1,0 +1,61 @@
+var React = require('react');
+
+var CreateNewClassComponent = React.createClass({
+    getInitialState: function() {
+        return {
+            newname: ''
+        }
+    },
+    handleChange: function(e) {
+        this.setState({newname: e.target.value});
+    },
+    handleSubmit: function(type) {
+        var is_event = false;
+        if (type == "event") {is_event = true;}
+        var newItem = this.state.newname;
+        if (NameCheck.check(newItem)) {
+            if (this.props.onSubmit(newItem, is_event)) {
+                this.setState({newname: ''});
+            }
+        }
+    },
+    submitData: function() {
+        this.handleSubmit("data");
+    },
+    submitEvent: function() {
+        this.handleSubmit("event");
+    },
+    handleEnterSubmit: function(e) {
+        if (e.keyCode == 13) {
+            this.submitData()
+        }
+    },
+    render: function() {
+        return (
+            <div className="panel panel-default">
+                <div className="panel-heading">
+                    Create a new class
+                </div>
+                <div className="panel-body">
+                    <div className="input-group">
+                        <input
+                            type="text"
+                            className="form-control"
+                            id="domainmodelDataClassName"
+                            placeholder="New class"
+                            value = {this.state.newname}
+                            onChange = {this.handleChange}
+                            onKeyDown = {this.handleEnterSubmit}
+                            />
+                        <div className="input-group-btn">
+                            <button className="btn btn-primary" onClick={this.submitData}>Create dataclass</button>
+                            <button className="btn btn-primary" onClick={this.submitEvent}>Create eventclass</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
+});
+
+module.exports = CreateNewClassComponent;
