+++ public_src/js/react/domainmodeleditor/DomainModelEditor.js
@@ -6,81 +6,6 @@ var Config = require('./../../config');
 var DataClassComponent = require('./DataClass');
 var OLCEditorComponent = require('./../olceditor');
 
-var CreateNewClassComponent = React.createClass({
-    getInitialState: function() {
-        return {
-            newname: ''
-        }
-    },
-    handleChange: function(e) {
-        this.setState({newname: e.target.value});
-    },
-    handleSubmit: function(type) {
-        var is_event = false;
-        if (type == "event") {is_event = true;}
-        var newItem = this.state.newname;
-        if (NameCheck.check(newItem)) {
-            if (this.props.onSubmit(newItem, is_event)) {
-                this.setState({newname: ''});
-            }
-        }
-    },
-    submitData: function() {
-        this.handleSubmit("data");
-    },
-    submitEvent: function() {
-        this.handleSubmit("event");
-    },
-    handleEnterSubmit: function(e) {
-        if (e.keyCode == 13) {
-            this.submitData()
-        }
-    },
-    render: function() {
-        return (
-            <div className="panel panel-default">
-                <div className="panel-heading">
-                    Create a new class
-                </div>
-                <div className="panel-body">
-                    <div className="input-group">
-                        <input
-                            type="text"
-                            className="form-control"
-                            id="domainmodelDataClassName"
-                            placeholder="New class"
-                            value = {this.state.newname}
-                            onChange = {this.handleChange}
-                            onKeyDown = {this.handleEnterSubmit}
-                            />
-                        <div className="input-group-btn">
-                            <button className="btn btn-primary" onClick={this.submitData}>Create dataclass</button>
-                            <button className="btn btn-primary" onClick={this.submitEvent}>Create eventclass</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        );
-    }
-});
-
-var OperationsComponent = React.createClass({
-    render: function() {
-        return (
-            <div className="panel panel-default">
-                <div className="panel-heading">
-                    Operations
-                </div>
-                <div className="panel-body">
-                    <div className="btn-group btn-block">
-                        <button className="btn btn-success" onClick={this.props.onSave}>Save</button>
-                    </div>
-                </div>
-            </div>
-        );
-    }
-});
-
 var DomainModelEditorComponent = React.createClass({
     getInitialState: function() {
         return {
@@ -257,7 +182,7 @@ var DomainModelEditorComponent = React.createClass({
                     </div>
                   </div>
                 </div>
-              </div>;
+              </div>
 
               </div>
         )
