+++ public_src/js/react/domainmodeleditor/Operations.js
@@ -1,0 +1,20 @@
+var React = require('react');
+
+var OperationsComponent = React.createClass({
+    render: function() {
+        return (
+            <div className="panel panel-default">
+                <div className="panel-heading">
+                    Operations
+                </div>
+                <div className="panel-body">
+                    <div className="btn-group btn-block">
+                        <button className="btn btn-success" onClick={this.props.onSave}>Save</button>
+                    </div>
+                </div>
+            </div>
+        );
+    }
+});
+
+module.exports = OperationsComponent;
